Intermediary Level Added and Bug Fix
Added an intermediate level introducing the idea of the platforms separately from the idea of the triangle to go in between the demo and Jesus' level.

Fixed a large bug where the trigger would allow the player to jump even if the platform was not there:

If anything goes wrong with your levels when you are building them, make sure that each platform either:
a) if it's the basic ground, it should be tagged "Sand"
b) all platforms should have the 'ground check' script attached to them
c) The inverse hero should be tagged "Inverse" and the main hero should be tagged "Player".